Create manual_boxplots_DSV_growing_zones.R
pixf is a function to get values() from rasters and create data frames for each pixel of day of each year. Large data frames are created in several cases. sumgraphf is the manual approach to a boxplot that is necessary since ggplot geom_boxplot can't handle all the rows. In essence, use summary() and use the mean, quartiles, etc, from that to create the boxplots manually.